оновлює функції

+++ controllers/notices/removeNoticeById.js
@@ -3,11 +3,13 @@ const {NotFound} = require('http-errors')
 
 const removeNoticeById = async (req, res) => {
   const { noticeId } = req.params;
-  console.log(req.params)
-    const notice = await Notice.findByIdAndRemove(noticeId)
-    if (!notice) {
-        throw new NotFound("Notice does not exist")
-    }
+  const { _id: owner } = req.user;
+  const notice = await Notice.findByIdAndRemove({ _id: noticeId, owner });
+
+  if (!notice) {
+    throw new NotFound("Notice does not exist");
+  };
+
   res.json({ message: "Notice has been removed " });
 };
